move 5 card hand rank to generic poker file
